Create SendableSpeedController
