+++ wpilibc/src/main/native/cpp/PWMSpeedController.cpp
@@ -1,13 +1,11 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2008-2018 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2008-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
 #include "frc/PWMSpeedController.h"
-
-#include "frc/smartdashboard/SendableBuilder.h"
 
 using namespace frc;
 
@@ -25,16 +23,8 @@ bool PWMSpeedController::GetInverted() const { return m_isInverted; }
 
 void PWMSpeedController::Disable() { SetDisabled(); }
 
-void PWMSpeedController::StopMotor() { PWM::StopMotor(); }
+void PWMSpeedController::StopMotor() { PWMBase::StopMotor(); }
 
 void PWMSpeedController::PIDWrite(double output) { Set(output); }
 
-PWMSpeedController::PWMSpeedController(int channel) : PWM(channel) {}
-
-void PWMSpeedController::InitSendable(SendableBuilder& builder) {
-  builder.SetSmartDashboardType("Speed Controller");
-  builder.SetActuator(true);
-  builder.SetSafeState([=]() { SetDisabled(); });
-  builder.AddDoubleProperty("Value", [=]() { return GetSpeed(); },
-                            [=](double value) { SetSpeed(value); });
-}
+PWMSpeedController::PWMSpeedController(int channel) : PWMBase(channel) {}
