+++ wpilibc/src/main/native/cpp/SpeedControllerGroup.cpp
@@ -61,11 +61,3 @@ void SpeedControllerGroup::StopMotor() {
 }
 
 void SpeedControllerGroup::PIDWrite(double output) { Set(output); }
-
-void SpeedControllerGroup::InitSendable(SendableBuilder& builder) {
-  builder.SetSmartDashboardType("Speed Controller");
-  builder.SetActuator(true);
-  builder.SetSafeState([=]() { StopMotor(); });
-  builder.AddDoubleProperty("Value", [=]() { return Get(); },
-                            [=](double value) { Set(value); });
-}
