+++ wpilibc/src/main/native/include/frc/NidecBrushless.h
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2019 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2017-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -11,8 +11,7 @@
 #include "frc/ErrorBase.h"
 #include "frc/MotorSafety.h"
 #include "frc/PWM.h"
-#include "frc/SpeedController.h"
-#include "frc/smartdashboard/Sendable.h"
+#include "frc/SendableSpeedController.h"
 #include "frc/smartdashboard/SendableHelper.h"
 
 namespace frc {
@@ -22,9 +21,8 @@ class SendableBuilder;
 /**
  * Nidec Brushless Motor.
  */
-class NidecBrushless : public SpeedController,
+class NidecBrushless : public SendableSpeedController,
                        public MotorSafety,
-                       public Sendable,
                        public SendableHelper<NidecBrushless> {
  public:
   /**
