+++ wpilibj/src/main/java/edu/wpi/first/wpilibj/NidecBrushless.java
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2019 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2017-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -15,7 +15,7 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableRegistry;
 /**
  * Nidec Brushless Motor.
  */
-public class NidecBrushless extends MotorSafety implements SpeedController, Sendable,
+public class NidecBrushless extends MotorSafety implements SendableSpeedController,
     AutoCloseable {
   private boolean m_isInverted;
   private final DigitalOutput m_dio;
