+++ wpilibj/src/main/java/edu/wpi/first/wpilibj/PWMSpeedController.java
@@ -1,5 +1,5 @@
 /*----------------------------------------------------------------------------*/
-/* Copyright (c) 2008-2018 FIRST. All Rights Reserved.                        */
+/* Copyright (c) 2008-2020 FIRST. All Rights Reserved.                        */
 /* Open Source Software - may be modified and shared by FRC teams. The code   */
 /* must be accompanied by the FIRST BSD license file in the root directory of */
 /* the project.                                                               */
@@ -7,12 +7,10 @@
 
 package edu.wpi.first.wpilibj;
 
-import edu.wpi.first.wpilibj.smartdashboard.SendableBuilder;
-
 /**
  * Common base class for all PWM Speed Controllers.
  */
-public abstract class PWMSpeedController extends PWM implements SpeedController {
+public abstract class PWMSpeedController extends PWM implements SendableSpeedController {
   private boolean m_isInverted;
 
   /**
@@ -77,13 +75,5 @@ public abstract class PWMSpeedController extends PWM implements SpeedController 
   @Override
   public void pidWrite(double output) {
     set(output);
-  }
-
-  @Override
-  public void initSendable(SendableBuilder builder) {
-    builder.setSmartDashboardType("Speed Controller");
-    builder.setActuator(true);
-    builder.setSafeState(this::setDisabled);
-    builder.addDoubleProperty("Value", this::getSpeed, this::setSpeed);
   }
 }
